working tests, linted
