+++ packages/frame-utils/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @turnkey/utils
+# @turnkey/frame-utils
 
 ## 0.1.0
