+++ packages/frame-utils/README.md
@@ -1,3 +1,3 @@
-# @turnkey/utils
+# @turnkey/frame-utils
 
 This package consolidates some common cryptographic utilities used across our applications, particularly primitives related to encoding/decoding.
