+++ packages/frame-utils/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@turnkey/utils",
+  "name": "@turnkey/frame-utils",
   "version": "0.1.0",
   "main": "./dist/index.js",
   "module": "./dist/index.mjs",
@@ -21,7 +21,7 @@
   "repository": {
     "type": "git",
     "url": "https://github.com/tkhq/sdk.git",
-    "directory": "packages/utils"
+    "directory": "packages/frame-utils"
   },
   "files": [
     "dist/",
